Added barcode

+++ src/actions/Item/deleteItem.ts
@@ -1,4 +1,5 @@
 "use server";
+import { del } from "@vercel/blob";
 import fs from "fs";
 import path from "path";
 import { deleteItemById } from "@/data/item";
@@ -7,9 +8,22 @@ export const deleteItem = async (id: number) => {
   try {
     const item = await deleteItemById(id);
     if (item.image) {
-      const fullImagePath = path.join(process.cwd(), "public", item.image);
-      if (fs.existsSync(fullImagePath)) {
-        fs.unlinkSync(fullImagePath);
+      if (process.env.MODE === "LOCAL") {
+        const fullImagePath = path.join(process.cwd(), "public", item.image);
+        const fullBarcodeImagePath = path.join(
+          process.cwd(),
+          "public",
+          item.barcodeImage,
+        );
+        if (fs.existsSync(fullImagePath)) {
+          fs.unlinkSync(fullImagePath);
+        }
+        if (fs.existsSync(fullBarcodeImagePath)) {
+          fs.unlinkSync(fullBarcodeImagePath);
+        }
+      } else if (process.env.MODE === "VERCEL") {
+        await del(item.image);
+        await del(item.barcodeImage);
       }
     }
     return { success: "Item deleted successfully" };
